+++ hiclass/__init__.py
@@ -9,6 +9,7 @@ from .MultiLabelLocalClassifierPerNode import MultiLabelLocalClassifierPerNode
 from .MultiLabelLocalClassifierPerParentNode import (
     MultiLabelLocalClassifierPerParentNode,
 )
+from .Explainer import Explainer
 from ._version import get_versions
 
 __version__ = get_versions()["version"]
